macro out the repeated code
