+++ Cargo.toml
@@ -1,9 +1,17 @@
 [package]
 name = "cli-rs"
-version = "0.1.7"
+version = "0.1.8"
 edition = "2021"
 license = "Unlicense"
 description = "create clis with sophisticated completions"
 
 [dependencies]
 colored = "2"
+cli-rs-command-gen = { path = "src/command/cli-rs-command-gen" }
+
+[workspace]
+resolver="2"
+members = [
+    "",
+    "src/command/cli-rs-command-gen"
+]
