+++ src/command/cli-rs-command-gen/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "cli-rs-command-gen"
+version = "0.1.8"
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = "2.0"
+quote = "1.0"
+proc-macro2 = "1.0"
